backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
mysql-connector-python==8.3.0
pydantic==2.7.1
passlib==1.7.4
python-jose==3.3.0